Major changes to source code, you may want to wipe and clone fresh.
Server system is now fully compliant with Hashtable based commands for network IO, separating it from the rest of the system.

Client connections have a custom 10 second timeout and are now multi-threaded, the GUI does not "lock-up" during connection requests.

Removed Debug class files, not enough time to implement a full testing environment, so we are aiming strait for a production model.

Revised the naming scheme to unify and simplify file identification.

Next update will be client and server implementations of ChatSystem and a ChatGUI for the client.  After that I will work on MatchSystem implementation for match-making.

Chat & Game GUI ideas, create a SystemGUI that loads after the Login, the SystemGUI displays a basic menu, and has the chat GUI inside of it.  Tabbed based Game menu extends to one side, allowing multiple games to be played by the user using JTabbedPanel or similar class.