Confirmed foreign character support in Chat system, adjusted on-send to clear the input text field.  Also, focus is set at startup to the input box.  Completed the Challenge operation in the Chat GUI.
Both client & server side Game Factories and Systems are now implemented, the Client side displays a 6 row by 8 column board on game instantiation, and the server prepares a set of random tiles.

Upcoming changes include making it a bit smaller to work in more screen resolutions, so a 4 row by 6 column game instance, and adding a score-board for the players and points system at the bottom.  The final step after that will be implementing the commands to display tiles, swap turns, increment points, and finish a game.