Fixed buggy score counting "ending" of games server-side.  Added a 2 second timeout after the second displayed icon that presents sending additional clicks, and hides all displayed icons after.  I then removed the hack that allowed clicking already displayed icons.
Implemented KILL command for disconnected users and potentially a quit option from the menu in the future.  Addressed a bug I missed previously where games were being removed from the arraylist but not the display.

At this stage we are done with functionality, anything else is aesthetics.

I will run a stress test, loading the Server on my home machine, and connecting as many instances and games as I can from the school computers.  If I can get 25+ users and over 100 games running that would be a good demonstration.

After that I will double check that the program meets the requirements for the project, rebuild the UML diagrams for Client and Server, and update the design document.