Updated the bash and batch scripts to generate JavaDocs pre-compile to a javadocs/ directory, which is next included with the jar file.
JavaDocs are not 100% complete, basic information exists for all classes, but individual methods do not have comments.  This is fine since most methods are inappropriately named anyways, I would rather address that first then add JavaDoc comments.  For now, it's on a "to do if we ever work on this in the future" list.