Integrated a sleep process into the UPDATE GameSystem interpreter, such that the tiles remain displayed for two seconds before being removed and the score updating.
Modified Game Instances to check isVisible instead of isEnabled for win condition.

Integrated POKE command on Server & Client GameSystems which will disable or enable all buttons to indicate whose turn it is visually.  Added "TURN" to NEWGAME command, Client side modified to run swapEnabled for first-turn on new game.

Added a 2 second delay to client side POKE interpreter, this is to accommodate the first aesthetic change and is a flawed solution.  If network IO is erratic and both the update and poke commands reach the opponents client at the same time, they have an opportunity to click one of the tiles that is about to be removed.  The tile won't exist on the server, so it won't wreck scoring, but the Server may throw an error since the tile with those coordinates will no-longer exist.